+++ tests/build-testing.R
@@ -1,0 +1,1 @@
+rhub::check_for_cran()